+++ .changeset/friendly-lamps-cheer.md
@@ -1,5 +1,5 @@
 ---
-'@keystone-next/keystone-legacy': patch
+'@keystone-next/keystone': patch
 ---
 
 Removed unused and inaccessible code from the core.
