+++ .changeset/friendly-rice-clap.md
@@ -2,4 +2,4 @@
 '@keystone-next/utils-legacy': minor
 ---
 
-Added a humanize fn which is string transformation logic ported over from @keystone-next/keystone-legacy.
+Added a humanize fn which is string transformation logic ported over from @keystone-next/keystone.
