+++ .changeset/large-beers-drive.md
@@ -30,7 +30,6 @@
 '@keystone-next/types': major
 '@keystone-next/access-control-legacy': major
 '@keystone-next/adapter-prisma-legacy': major
-'@keystone-next/keystone-legacy': major
 '@keystone-next/server-side-graphql-client-legacy': major
 '@keystone-next/test-utils-legacy': major
 '@keystone-next/utils-legacy': major
