+++ .changeset/nice-bugs-repeat.md
@@ -1,5 +1,5 @@
 ---
-'@keystone-next/keystone-legacy': patch
+'@keystone-next/keystone': patch
 ---
 
 Moved core logic in keyToLabel to the @keystone-next/utils-legacy package.
