+++ .changeset/proud-apricots-punch.md
@@ -1,5 +1,5 @@
 ---
-'@keystone-next/keystone-legacy': major
+'@keystone-next/keystone': patch
 ---
 
 Removed legacy support for auxilliary lists.
