+++ .changeset/real-news-reflect.md
@@ -1,0 +1,6 @@
+---
+'@keystone-next/keystone': patch
+'@keystone-next/adapter-prisma-legacy': patch
+---
+
+Consolidated the core code from the `@keystone-next/keystone-legacy` package into `@keystone-next/keystone`.
