+++ packages-next/keystone/package.json
@@ -27,12 +27,13 @@
     "@graphql-tools/merge": "^6.2.11",
     "@graphql-tools/schema": "^7.1.3",
     "@hapi/iron": "^6.0.0",
-    "@keystone-next/adapter-prisma-legacy": "5.0.0",
+    "@keystone-next/access-control-legacy": "^9.0.1",
+    "@keystone-next/adapter-prisma-legacy": "^5.0.0",
     "@keystone-next/admin-ui": "^13.0.0",
     "@keystone-next/fields": "^6.0.0",
-    "@keystone-next/keystone-legacy": "^23.0.0",
-    "@keystone-next/server-side-graphql-client-legacy": "3.0.1",
+    "@keystone-next/server-side-graphql-client-legacy": "^3.0.1",
     "@keystone-next/types": "^16.0.0",
+    "@keystone-next/utils-legacy": "^8.0.0",
     "@preconstruct/next": "^2.0.0",
     "@prisma/client": "2.20.1",
     "@prisma/migrate": "2.20.2",
@@ -49,16 +50,19 @@
     "@types/prompts": "^2.0.10",
     "@types/source-map-support": "^0.5.3",
     "@types/uid-safe": "^2.1.2",
+    "apollo-errors": "^1.9.0",
     "apollo-server-express": "^2.22.2",
     "apollo-server-micro": "^2.22.2",
     "apollo-server-types": "^0.7.0",
     "chalk": "^4.1.0",
     "cookie": "^0.4.1",
     "cors": "^2.8.5",
+    "cuid": "^2.1.8",
     "execa": "^5.0.0",
     "express": "^4.17.1",
     "fs-extra": "^9.1.0",
     "graphql": "^15.5.0",
+    "graphql-type-json": "^0.3.2",
     "graphql-upload": "^11.0.0",
     "meow": "^9.0.0",
     "next": "^10.0.9",
@@ -70,7 +74,9 @@
     "prompts": "^2.4.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "serialize-error": "^8.0.1",
     "source-map-support": "^0.5.19",
+    "stack-utils": "^2.0.3",
     "typescript": "^4.2.3",
     "uid-safe": "^2.1.5"
   },
