+++ packages-next/keystone/src/lib/core/Keystone/format-error.d.ts
@@ -1,0 +1,1 @@
+export const formatError: any;
