+++ packages-next/keystone/src/lib/core/Keystone/index.d.ts
@@ -1,0 +1,1 @@
+export const Keystone: any;
