+++ packages-next/keystone/src/lib/createApolloServer.ts
@@ -3,11 +3,9 @@ import { GraphQLSchema } from 'graphql';
 import { ApolloServer as ApolloServerMicro } from 'apollo-server-micro';
 import { ApolloServer as ApolloServerExpress } from 'apollo-server-express';
 import type { Config } from 'apollo-server-express';
-
-// @ts-ignore
-import { formatError } from '@keystone-next/keystone-legacy/lib/Keystone/format-error';
 import type { CreateContext, SessionStrategy } from '@keystone-next/types';
 import { createSessionContext } from '../session';
+import { formatError } from './core/Keystone/format-error';
 
 export const createApolloServerMicro = ({
   graphQLSchema,
