--- packages/keystone/.npmignore
@@ -1,2 +1,0 @@
-**/*.md
-**/*.test.js
