--- packages/keystone/index.js
@@ -1,2 +1,0 @@
-const Keystone = require('./lib/Keystone');
-module.exports = { Keystone };
