--- packages/keystone/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@keystone-next/keystone-legacy",
-  "description": "The main @keystonejs class & CLI. This is where the magic happens.",
-  "version": "23.0.0",
-  "author": "The KeystoneJS Development Team",
-  "license": "MIT",
-  "engines": {
-    "node": ">=12.13.0"
-  },
-  "dependencies": {
-    "@keystone-next/access-control-legacy": "^9.0.1",
-    "@keystone-next/utils-legacy": "^8.0.0",
-    "apollo-errors": "^1.9.0",
-    "apollo-server-express": "^2.22.2",
-    "cuid": "^2.1.8",
-    "ensure-error": "^3.0.1",
-    "express": "^4.17.1",
-    "graphql-type-json": "^0.3.2",
-    "graphql-upload": "^11.0.0",
-    "pluralize": "^8.0.0",
-    "serialize-error": "^8.0.1",
-    "stack-utils": "^2.0.3"
-  },
-  "devDependencies": {
-    "@keystone-next/fields": "^6.0.0",
-    "@keystone-next/test-utils-legacy": "*",
-    "graphql": "^15.5.0"
-  },
-  "repository": "https://github.com/keystonejs/keystone/tree/master/packages/keystone"
-}
