+++ tests/api-tests/fields/unsupported.test.ts
@@ -2,6 +2,7 @@ import path from 'path';
 import globby from 'globby';
 import { multiAdapterRunners, setupFromConfig, testConfig } from '@keystone-next/test-utils-legacy';
 import { createSchema, list } from '@keystone-next/keystone/schema';
+import { BaseKeystone } from '@keystone-next/types';
 
 const testModules = globby.sync(`{packages,packages-next}/**/src/**/test-fixtures.{js,ts}`, {
   absolute: true,
@@ -29,7 +30,7 @@ multiAdapterRunners().map(({ provider, after }) => {
                 await mod.afterAll();
               }
               // We expect setup to fail, so disconnect can be a noop
-              await after({ disconnect: () => {} });
+              await after({ disconnect: async () => {} } as BaseKeystone);
             });
 
             test('Throws', async () => {
